+++ vercel.json
@@ -3,5 +3,9 @@
   "builds": [{"src": "main.py", "use": "@vercel/python"}],
   "routes": [
     {"src":  "/(.*)", "dest":  "main.py"}
-  ]
+  ],
+  "crons": [{
+    "path": "/crons/notify",
+    "schedule": "*/5 * * * *"
+  }]
 }
